+++ src/assets/InputAsset.php
@@ -139,6 +139,7 @@ class InputAsset extends FieldAsset
             'maxBlocks' => $field->maxBlocks,
             'maxTopBlocks' => $field->maxTopBlocks,
             'maxLevels' => (int)$field->maxLevels,
+            'showBlockTypeHandles' => Craft::$app->getUser()->getIdentity()->getPreference('showFieldHandles'),
         ];
 
         $encodedJsSettings = Json::encode($jsSettings, JSON_UNESCAPED_UNICODE);